feat: add recent games category lane
