+++ src/App.tsx
@@ -21,7 +21,7 @@ const Header = lazy(() => import('./components/UI/Header'))
 function App() {
   const context = useContext(ContextProvider)
   const user = configStore.get('userInfo')
-  const { data: library, refresh } = context
+  const { data: library, refresh, recentGames } = context
 
   const dlcCount = library.filter((lib) => lib.install.is_dlc)
   const numberOfGames = library.length - dlcCount.length
@@ -38,6 +38,7 @@ function App() {
                   renderBackButton={false}
                   numberOfGames={numberOfGames}
                 />
+                {!!recentGames.length && <Library showRecentsOnly library={recentGames} />}
                 <Library library={library} />
               </>
                 : <Login refresh={refresh} />}
